changed casts again

+++ src/ssimCheck.cpp
@@ -214,7 +214,7 @@ int main(int argc, char **argv){
 		tests = "PSNR and RMSE";
 	}
 
-	string sourceDims = std::to_string(refS.width) + "x" + std::to_string(refS.height);
+	string sourceDims = std::to_string(static_cast<int>(refS.width)) + "x" + std::to_string(static_cast<int>(refS.height));
 	string testDims = std::to_string(dstS.width) + "x" + std::to_string(dstS.height);
 	// write json header to output file
 	fprintf(outputfile, templateHeader, referenceVideo, sourceDims.c_str(), totalSrcFrames, testVideo, testDims.c_str(), totalTstFrames, tests.c_str());
